backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
agno>=2.5.0
httpx==0.27.0
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
instagrapi>=2.1.0
pydantic-settings==2.5.0
python-dotenv==1.0.1
groq>=0.11.0
openai>=1.60.0
anthropic>=0.36.0
bcrypt>=4.1.0
PyJWT>=2.8.0
